Fix template finding logic
For the moment we encourage the use of Rabl and ERb. The former can
render multiple formats, so including the format in the path doesn't
make sense. E.g., `user.rabl` could render in both JSON and XML, so we
wouldn't want to have both `user.json.rabl` and `user.xml.rabl`
templates.

Our current path query is:

    {.json,.*}{.rabl,.erb,}

The problem with this is a request for JSON could render a Haml template
if it exists and has either `.haml` or `.html.haml` as the suffix
(because it will match first).

The solution is to favor templates with the format listed, otherwise
falling back to Rabl (because it can render multiple formats) or ERb
(because it can conventionally be used ambiguously in Ruby):

    .{json.{rabl,erb,*},{rabl,erb,*}}

Signed-off-by: Stephen Celis <stephen@stephencelis.com>