Don't define OPTIONS route for paths that do not define methods
This fixes a bug where mounting another Crepe::API app at a path
produces an allowed array of `[nil]` which throws an error on
`#sort!`.

If a path doesn't have any methods defined, it's because the endpoint
wants to handle routing itself (ie. it's probably another Rack app),
so we want to delegate generation of OPTIONS routes for that path to
the Rack app if it decides to.

Signed-off-by: Evan Owen <kainosnoema@gmail.com>