Raise exception when nesting middleware
Middleware is mounted, at the moment, before an API's router. Routers,
meanwhile, map out an API's mounted endpoints and Rack applications:

                                             #<Crepe::Endpoint>
  Crepe::API (Shared Middleware + Router) -> #<Crepe::Endpoint>
                                             #<Proc:(env)>

To allow for nested middleware, it would need to be mounted between the
Crepe::API router and each endpoint:

                         Isolated Middleware   -> #<Crepe::Endpoint>
  Crepe::API (Router) -> Duplicated Middleware -> #<Crepe::Endpoint>
                         Unique Middleware     -> #<Proc:(env)>

The above model is more complex and makes every endpoint heavier.

Crepe allows for you to mount an API inside an API (for organization,
etc.):

                                  Crepe::API (Middleware - Root Middleware)
  Crepe::API (Root Middleware) -> Crepe::API (Middleware - Root Middleware)
                                  Crepe::API (Middleware - Root Middleware)

Because of the above ability, you can scope the use of middleware to a
specific subsection of your API. Crepe will automatically remove
duplicate middleware in the stack when the application is run.

It's confusing, though, to be able to declare `use Middleware` within a
namespace and have it apply universally within that API (and its mounted
apps), so we should probably raise an exception here.

In the future, it may be desirable to scope middleware within
namespaces, but the complexities of inner-mounted APIs (and other
applications) must be dealt with first.

Signed-off-by: Stephen Celis <stephen@stephencelis.com>