Store rescuers in a hash
Storing rescuers in an array didn't make a lot of sense. Duplicates
would be appended and silently ignored. It also complicated lookup. By
using a hash instead, we make it easy to override rescuers in nested
scopes, and by fetching all ancestral matches, we can select the most
specific rescuer for the exception raised.

Signed-off-by: Stephen Celis <stephen@stephencelis.com>