RESTful status updates
Deleting the body of DELETE requests was a mistake. There are perfectly
valid reasons to return content. Meanwhile, it's perfectly acceptable to
return empty bodies for POST, PUT, and PATCH requests. Because of this,
let's return 201 for POSTs that return content, 200s for the rest, and
if no content is returned, 204s.

Signed-off-by: Stephen Celis <stephen@stephencelis.com>