Add stricter, declarative version configuration
Our existing solution relied on Middleware::ContentNegotiation,
which was based on another middleware that mutated PATH_INFO by
appending the path extension. We prepended PATH_INFO with any
version found in the Accept header, and appended with the first
format.

There were multiple issues with this approach.

By checking acceptance in middleware before the route set, it took
the first match and passed it through, even if the endpoint didn't
support it. The endpoint could, however, support another MIME type
listed later in the Accept header. This logic needs to live in the
endpoint itself. E.g., WebKit sends:

  text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8

Crepe APIs couldn't load in the browser because the format was
set to HTML and a 406 Not Acceptable was returned instead.

This also prevented us from defining versions within namespaces.

API.version will now require a choice, so that we can best use
Rack::Mount to direct requests to their intended endpoint.

We've removed Middleware::ContentNegotiation for the moment, though
we need to bring some of its logic back in, especially that which
sets env['crepe.vendor'], etc.

Signed-off-by: Stephen Celis <stephen.celis@gmail.com>
Signed-off-by: Evan Owen <kainosnoema@gmail.com>