Add Crepe.env and Crepe.root
Crepe.env looks for CREPE_ENV, RACK_ENV, or assumes development.

Crepe.root assumes PWD. We could filter the call stack instead (as Rails
and Padrino do), but in favor of simplicity (and the fact that most
scripts and binaries depend on PWD to function properly), let's avoid
that for now.

Signed-off-by: Evan Owen <kainosnoema@gmail.com>
Signed-off-by: Stephen Celis <stephen@stephencelis.com>