Update README with example (and Code Climate)
This is a real-worldish example that demonstrates a bunch of things
Crepe does, but it's a bit messy (as real-world APIs tend to be). We
should add some more examples with best/better practices, and maybe
ditch this one entirely in the future. But hey, it's better than it was.

Signed-off-by: Stephen Celis <stephen@stephencelis.com>