Fix Crepe::Params#respond_to?
Crepe::Params instances respond to more than we're saying. In
method_missing, we'll proxy any method call through to the underlying
HashWithIndifferentAccess. In respond_to?, we claim that a Crepe::Params
instance will only respond to require, permit, permitted?, permit!, and
instance methods present on BasicObject. We should change the `super`
check to a check that the underlying HashWithIndifferentAccess responds
to the method.

Signed-off-by: David Celis <me@davidcel.is>