Rewrite README
The current README is a bit difficult to parse: it presents nothing more
than a verbose example that is somewhat terse and not very explanatory.
In order to draw more users, we should present a clean example that
showcases Crepe's basic, elegant features. Meanwhile, I will be adding
pages to the wiki that show off Crepe's more advanced features, such as
HEAD/OPTIONS routing, streaming, and configuration.

Signed-off-by: David Celis <me@davidcel.is>